made ShareFolderLink class
I needed a different data structure and class to insert and keep track of share folder links that is more flat and does not group by folder name.
This way when the client wants to get the folders that are shared from other companies...it can just filter this list by sharedCompanyId and get the folderNames. I added permissions just incase I need it.