made ShareFolder class
This is the class that groups permissions by foldername and originCompanyId.
This is a good data structure that will be serialized to json to send to the client and a good candidate for the shared folders management list where you can do updates on the permissions and companies that folder is shared with.